requirements.txt
# Polymarket API
requests==2.31.0
aiohttp==3.9.1

# Blockchain (for verification)
web3==6.11.3

# Database
psycopg[binary]>=3.1.0
sqlalchemy==2.0.23
alembic==1.13.1

# OSINT
beautifulsoup4==4.12.2
selenium==4.16.0

# Analysis
pandas==2.1.4
numpy==1.26.2

# Alerts
python-telegram-bot==20.7
python-dotenv==1.0.0

# Scheduling
apscheduler==3.10.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Dashboard
fastapi==0.108.0
uvicorn==0.25.0
streamlit==1.29.0
plotly==5.18.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
